# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: StudyForge AI API
  version: 0.1.0
  summary: Gateway for the StudyForge AI platform.
  description: |
    Conventions (mandatory):
      - All errors are application/problem+json (RFC 9457) with a stable `code`.
      - Lists paginate by opaque cursor only.
      - State-changing endpoints require `Idempotency-Key`.
      - Rate limits exposed via `RateLimit-*` headers.
      - Streams are SSE; bidirectional surfaces use WebSocket.
    See docs/architecture/04-api-design.md.
  contact:
    name: StudyForge Platform
    url: https://studyforge.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: Local dev
  - url: https://api.studyforge.ai
    description: Production
...
security:
  - bearer: []
paths:
...
  /v1/uploads/init:
    post:
      tags: [uploads]
      summary: Reserve + sign upload URL
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadInitRequest" }
      responses:
        "200":
          description: Signed URL + uploadId
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadInitResponse" }
        "413": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /v1/uploads/{id}/complete:
    post:
      tags: [uploads]
      summary: Mark upload complete, enqueue ingest
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadBatch" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 64 }
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Problem:
      description: Problem (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "Stable machine-readable code (e.g. upload.size-exceeded)" }
        traceId: { type: string }
        tenantId: { type: string, format: uuid, nullable: true }
        fields:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              reason: { type: string }
...
    UploadInitRequest:
      type: object
      required: [filename, mime, sizeBytes, sha256]
      properties:
        courseId: { type: string, format: uuid, nullable: true }
        filename: { type: string, minLength: 1, maxLength: 255 }
        mime: { type: string }
        sizeBytes: { type: integer, minimum: 1, maximum: 2147483648 }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
    UploadInitResponse:
      type: object
      required: [uploadId, signedUrl, expiresAt]
      properties:
        uploadId: { type: string, format: uuid }
        signedUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        headers:
          type: object
          description: Headers the client MUST include with the PUT to S3.
    UploadBatch:
      type: object
      properties:
        id: { type: string, format: uuid }
        state: { type: string, enum: [initiated, uploading, uploaded, scanning, extracting, embedding, ready, failed, quarantined] }
        safetyFlags: { type: array, items: { type: string } }
        errorReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...